requirements.txt
asgiref==3.8.1
black==24.8.0
certifi==2024.7.4
click==8.1.7
distlib==0.3.8
Django==5.1
django-cors-headers==4.4.0
django-debug-toolbar==3.2.2
django-environ==0.11.2
django-linear-migrations==2.13.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
setuptools==74.0.0
sqlparse==0.5.1
virtualenv==20.26.3
